Proof of concept generator in Bison and Flex
Since all of the logic is duplicated I thought I should find a way to
automate the code generation. Here I present an experimental version of
the library generator, that's yet to be included in the building
process, where I experiment with Bison and Flex, by parsing a
configuration file and generating template and function code
automatically.

Documentations will be added after intervace and tool have been refined.